third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGList-parse-invalid-clears-items-expected.txt

This is a testharness.js-based test.
[FAIL] SVGLengthList clears items when attribute value is invalid
  assert_equals: SVGLengthList: partially invalid attribute value should result in 0 items expected 0 but got 2
[FAIL] SVGNumberList clears items when attribute value is invalid
  assert_equals: SVGNumberList: partially invalid attribute value should result in 0 items expected 0 but got 2
[FAIL] SVGPointList clears items when attribute value is invalid
  assert_equals: SVGPointList: partially invalid attribute value should result in 0 items expected 0 but got 2
Harness: the test ran to completion.
